Standarise entity migration metadata via trait
